backend/requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
Deprecated==1.3.1
dlib @ file:///C:/Users/lapx/Desktop/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.135.2
fastapi-mail==1.6.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
numpy==1.26.4
opencv-python==4.8.0.76
packaging==26.0
passlib==1.7.4
pillow==12.1.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
regex==2025.11.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
wrapt==2.1.2
